+++ test/extension-tests/misc/utils.js
@@ -16,34 +16,11 @@
  */
 
 import assert from "assert";
-import fs from "fs";
-import path from "path";
-import url from "url";
-
-/*
- * Standard-compliant polyfill for WebDriver#executeScript,
- * working around limitations of ChromeDriver <77,
- * enabling scripts to return a promise.
- */
-export async function executeScriptCompliant(driver, script, ...args) {
-  let [isError, value] = await driver.executeAsyncScript(`
-    let promise = (async function() { ${script} }).apply(null, arguments[0]);
-    let callback = arguments[arguments.length - 1];
-    promise.then(
-      res => callback([false, res]),
-      err => callback([true, err instanceof Error ? err.message : err])
-    );`, args);
-
-  if (isError)
-    throw new Error(value);
-  return value;
-}
 
 export async function checkLastError(driver, handle) {
   await driver.switchTo().window(handle);
 
-  let error = await executeScriptCompliant(
-    driver,
+  let error = await driver.executeScript(
     "return browser.runtime.sendMessage({type: \"debug.getLastError\"});"
   );
   if (error != null)
@@ -59,16 +36,4 @@ export async function runWithHandle(driver, handle, callback) {
   finally {
     await driver.switchTo().window(currentHandle);
   }
-}
-
-export async function loadModules(dirname) {
-  let modules = [];
-  for (let dir of await fs.promises.readdir(dirname, {withFileTypes: true})) {
-    let filename = path.resolve(dirname, dir.name);
-    let basename = path.parse(dir.name).name;
-    if (dir.isDirectory())
-      filename = path.join(filename, "index.js");
-    modules.push([await import(url.pathToFileURL(filename)), basename]);
-  }
-  return modules;
 }