+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2022-08-18
+
+- Refactor extension tests.
+
 ## 2022-08-11
 
 - Add iframe exception testpage to index page.